Connection lost bug
Gametype: It's in the hubs.
Issue: Random connection lost
How to replicate: Basically, whenever you go to a hub. It sometimes give you an error of a connection lost. Here is an example:
Example: http://gyazo.com/b76e045680f8cccce3223bb3c3c5206a
When I go to the hub by using the netherstar, it gives me the error that's in the Gyazo screenshot.
Misc: